requirements.txt
python-telegram-bot==20.7
tinkoff-investments==0.2.0b59
python-dotenv
